+++ semana19/arquitetura-de-software/src/business/entities/task.ts
