+++ semana19/arquitetura-de-software/src/business/entities/user.ts
@@ -11,7 +11,6 @@ export type authenticationData = {
 export type user = {
    id: string,
    name: string,
-   nickname: string,
    email: string,
    password: string,
    role: USER_ROLES
