+++ semana19/arquitetura-de-software/src/business/services/hashManager.ts
