+++ semana19/arquitetura-de-software/src/business/services/idGenerator.ts
